+++ lib/features/media_conversion/convert_button.dart
@@ -6,9 +6,11 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:verticrowlabs_media_converter/features/file_parsing/file_parsing_barrel.dart';
 import 'package:verticrowlabs_media_converter/features/media_conversion/media_conversion_barrel.dart';
-import 'package:verticrowlabs_media_converter/infrastructure/common_variables/common_variables.dart';
+import 'package:verticrowlabs_media_converter/features/media_snipping/time_range_selector.dart';
+import 'package:verticrowlabs_media_converter/infrastructure/common_variables/common_enums.dart';
 import 'package:verticrowlabs_media_converter/infrastructure/common_variables/ffmpeg_cmd.dart';
 import 'package:verticrowlabs_media_converter/infrastructure/common_variables/pwsh_cmd.dart';
+import 'package:verticrowlabs_media_converter/infrastructure/models/mediatime.dart';
 
 ///[ConsumerWidget] Button that starts media conversion when pressed.
 ///disabled/enabled based on [buttonEnabled] value
@@ -82,6 +84,8 @@ Future<void> _convertMedia(WidgetRef ref) async {
         .read(conversionStatusProvider.notifier)
         .update((state) => ConversionStatus.done);
     ref.read(fileNameProvider.notifier).update((state) => '');
+    ref.read(startRangeProvider.notifier).update((state) => 0.0);
+    ref.read(maxTimeProvider.notifier).update((state) => '');
 
     //log('Finished');
   } else {
