+++ lib/features/media_conversion/media_resolution.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:verticrowlabs_media_converter/infrastructure/common_variables/common_variables.dart';
+import 'package:verticrowlabs_media_converter/infrastructure/common_variables/common_enums.dart';
 
 ///Creator that returns the chosen resolution as a MediaScale enum of either
 ///low, medium, high.
