+++ lib/infrastructure/common_variables/common_enums.dart
