+++ lib/infrastructure/common_widgets/file_selector_widget.dart
@@ -5,8 +5,8 @@ import 'package:verticrowlabs_media_converter/features/file_parsing/file_parsing
 import 'package:verticrowlabs_media_converter/features/media_conversion/conversion_status.dart';
 import 'package:verticrowlabs_media_converter/features/media_snipping/media_snipping.dart';
 import 'package:verticrowlabs_media_converter/features/thumbnail_generator/thumbnail_barrel.dart';
-import 'package:verticrowlabs_media_converter/infrastructure/common_variables/common_variables.dart';
-import 'package:verticrowlabs_media_converter/infrastructure/common_widgets/time_selector.dart';
+import 'package:verticrowlabs_media_converter/infrastructure/common_variables/common_enums.dart';
+import 'package:verticrowlabs_media_converter/features/media_snipping/time_range_selector.dart';
 
 ///[ConsumerWidget] that sets [fileInputStringProvider] when button is pressed.
 class FileSelector extends ConsumerWidget {
