+++ lib/infrastructure/pages/installer_page.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:verticrowlabs_media_converter/features/install_ffmpeg/install_ffmpeg.dart';
 import 'package:verticrowlabs_media_converter/features/install_ffmpeg/verify_install.dart';
-import 'package:verticrowlabs_media_converter/infrastructure/common_variables/common_variables.dart';
+import 'package:verticrowlabs_media_converter/infrastructure/common_variables/common_enums.dart';
 
 ///Screen that is displayed if ffmpeg is not installed
 class InstallerPage extends ConsumerWidget {
